@@ -1,43 +1,27 @@
 <template>
-  <div id="app">
-    <v-app>
-      <v-content v-if="loggedIn===null">
-        <v-container fill-height>
-          <v-layout align-center justify-center>
-            <v-flex>
-              <div class="text-xs-center">
-                <div class="headline my-5">Loading...</div>
-                <v-progress-circular size="100" indeterminate color="primary"></v-progress-circular>
-              </div>
-            </v-flex>
-          </v-layout>
-        </v-container>
-      </v-content>
-      <router-view v-else />
-      <NotificationsManager></NotificationsManager>
-    </v-app>
-  </div>
+  <img alt="Vue logo" src="./assets/logo.png" />
+  <HelloWorld msg="Hello Vue 3 + TypeScript + Vite" />
 </template>
 
 <script lang="ts">
-import { Component, Vue } from 'vue-property-decorator';
-import NotificationsManager from '@/components/NotificationsManager.vue';
-import { readIsLoggedIn } from '@/store/main/getters';
-import { dispatchCheckLoggedIn } from '@/store/main/actions';
+import { defineComponent } from 'vue'
+import HelloWorld from './components/HelloWorld.vue'
 
-@Component({
+export default defineComponent({
+  name: 'App',
   components: {
-    NotificationsManager,
-  },
+    HelloWorld
+  }
 })
-export default class App extends Vue {
-
-  get loggedIn() {
-    return readIsLoggedIn(this.$store);
-  }
-
-  public async created() {
-    await dispatchCheckLoggedIn(this.$store);
-  }
-}
 </script>
+
+<style>
+#app {
+  font-family: Avenir, Helvetica, Arial, sans-serif;
+  -webkit-font-smoothing: antialiased;
+  -moz-osx-font-smoothing: grayscale;
+  text-align: center;
+  color: #2c3e50;
+  margin-top: 60px;
+}
+</style>
